refactor: move FormState type to common for better reusability across sections
