+++ src/app/UI/admin/Forecasts/ForecastModal/ForecastForm/ProblemsSection/ProblemList/ProblemList.tsx
@@ -4,7 +4,7 @@ import { useTranslations } from 'next-intl'
 import { ProblemForm, type ProblemFormProps } from '../ProblemForm'
 import { ProblemItem } from './ProblemItem'
 
-import type { FormState } from '../ProblemsSection'
+import type { FormState } from '../../common'
 import type { Problem } from '@/business/types'
 
 type ProblemsListProps = {
