+++ src/app/UI/admin/Forecasts/ForecastModal/ForecastForm/common/index.ts
@@ -4,4 +4,6 @@ export { Footer } from './Footer'
 
 export { default as InputBlock } from './InputBlock'
 
+export type { FormState } from './types'
+
 export * from './hooks'
