+++ src/app/UI/admin/Forecasts/ForecastModal/ForecastForm/common/types.ts
@@ -1,0 +1,4 @@
+export type FormState = {
+  id?: string
+  mode: 'create' | 'edit'
+} | null
